git-svn-id: https://svn.code.sf.net/p/auber/code/branches/tulip_cpack@3284 f5d4c5f9-c943-4916-8a57-bd8605bf961c

+++ library/tulip-ogl/src/GlQuantitativeAxis.cpp
@@ -88,11 +88,11 @@ void GlQuantitativeAxis::buildAxisGraduations() {
 		maxV = max;
 	} else {
 		if (min >= 1) {
-			minV = minLog = log(min) / log(logBase);
-			maxV = maxLog = log(max) / log(logBase);
+			minV = minLog = log(min) / log((double)logBase);
+			maxV = maxLog = log(max) / log((double)logBase);
 		} else {
 			minV = minLog = 0;
-			maxV = maxLog = log(max - min) / log(logBase);
+			maxV = maxLog = log(max - min) / log((double)logBase);
 		}
 	}
 
@@ -169,7 +169,7 @@ Coord GlQuantitativeAxis::getAxisPointCoordForValue(double value) const {
 		if (min < 1) {
 			val += (1 - min);
 		}
-		val = log(val) / log(logBase);
+		val = log(val) / log((double)logBase);
 	}
 	if (ascendingOrder) {
 		 offset = (val - minV) * scale;
